requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
blinker==1.4
build==0.3.1.post1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-WTF==0.14.3
Flask-SSLify==0.1.5
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.9
pep517==0.10.0
pluggy==0.13.1
prodict==0.8.11
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-box==5.2.0
python-dateutil==1.5
python-dotenv==0.21.0
requests==2.25.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
